Split page-related code into page.py and some shared file code into files.py
This is a fast and dirty split, just looking at what functions/globals have to be moved into each file based on dependencies.  I haven't looked to see if other functions are also appropriate to be moved.  Nor have I tried to rewrite the code to minimize use of globals.