Don't italicize spp., ssp., or var. in the autocomplete box
The italicization and highlighting code is rewritten to be more generic and intuitive.

Interleaving of tags may have had problems in some browsers before, although it wasn't obvious in Chrome.  With the new changes, it was definitely a problem in Chrome, so I fixed the code to properly un-nest and re-nest interleaved tags.