Change javascripts to asynchronous
Search functionality is triggered as soon as both scripts are done loading.  This *greatly* improves search responsiveness.

The scroll position is restored repeatedly from the moment that the page becomes 'interactive' until it is 'complete'.  This fixes a bug in which a slow load/render would cause us to try to restore the scroll position before the page is rendered, which fail.
 The repeated restoration doesn't seem to improve responsiveness, possibly because offline testing is CPU limited, not network limited.