Added an icon in the upper left that links to the home page.
Now that a link to the home page is always visible in the header, the footer doesn't need to always be visible.
Moved the footer to be at the bottom of the scrollable page instead of locked to the bottom of the viewport.

Moved the contact info to the home page with an anchored (and highlighted) target.