Added more quota protection and related tasks
Added margin to usage while cache is being written.
Margin is deleted when done or on an error, making it more likely that space is available for a newly registering service worker.

Rearranged quota protection to have a more natural flow.
Added quota protection to the the margin write.
The margin is deleted prior to the final indexedDB update, so it shouldn't need quota protection.

Added code to ensure that an update is safely stopped before deleting all offline data.