Added support for 'x:ba' and similar flags at any taxon level
Unlike the explicit propagation of the 'complete' property (which I haven't yet removed), the 'x' value is implicitly propagated when the result is clear, and is not propagated when the result is unclear.

This fixes propagation of completeness into some unranked taxons (which don't support property propagation) and also identified a taxon with an unclear completeness (which I manually fixed).