Changed all cases of "<Genus> spp." to "genus <Genus>"
The Javascript search allows either style.
Also updated the Javascript search to prioritize a match on a scientific name without the preceding rank the same as a match from the start of the string.  So, e.g. a search for a genus name will list the genus page before all the species pages.