Added autofocus to the search bar and reverted to default scrolling of the entire page
I don't know if mobile browsers have changed the autofocus behavior so that the virtual keyboard doesn't automatically come up any more or if that was always the case, but I was doing it wrong before.  In any case, it now does what I want which is to allow typing right away on the desktop, but not popping up a virtual keyboard on mobile until the user taps the search box.

Since this change touched every HTML file, I also took this opportunity to revert to the default scroll behavior.  I.e. the scroll bar extends past the search field even though that part it isn't scrolled.  It's debatable whether this looks more or less clean, but at least the browser reliably restores this scroll position when navigating through the history, unlike the custom interior scrollable div.

This change also includes the beginnings of a key for manzanitas.