remove background margins
Now that there is no hidden page to scroll, Firefox no longer hijacks a drag to hide or reveal the URL bar, so I don't have to worry about revealing window parts behind the e_bg element.  Therefore I removed the 1000px margins from the element.

However, in case there is some other way to trigger a similar bug in any browser, I moved the background color to the body in the hopes that it might help.