Rearranged the gallery.html? search term to use the photo path instead of the page name and hash number
Other related changes allow the gallery to easily be invoked for a set of thumbnails *or* for an individual photograph or figure on any page.
Any link can go to the photo gallery, not just a thumbnail image.
Thumbnails are automatically generated (and used) for JPG figures.
Fixed the bay-area.jpg link to work when JavaScript is disabled.