added non-toxic plant info
Extensive rewrite of toxicity propagation to allow a taxon to have different ratings for different details.
The change in propagation doesn't just change the directly affected taxons, but I hope that it is smarter overall.
Aliasing is more sophisticated now since the toxic vs. non-toxic lists exposed a lot of ambiguous names that conflict with each other.