Don't rewrite HTML files if they are unchanged
HTML files are now written with "UNIX" \n line endings. This can slightly confuse GitHub desktop since git converts the HTML files to have WIndows-style \r\n line endings as they are fetched from the repository.  GitHub desktop thus says that all of each file is modified, but there are no differences in the file.  Note that after a commit, the history shows that the files with only line-ending differences were not actually modified, but GitHub desktop settles down and understands that the existing file is no different than the checked-in file.

Summary: after a fresh repository fetch, the script will re-write the HTML files (because the timestamps are different), and GitHub will think that these are changes, but it won't actually change the repository since the change is immaterial.