Allow prefix to be included in search
During regular search, the prefix invokes advanced search (if it isn't already a term exclusive to advanced search).
I.e. "within <taxon>" does an advanced search instead of going to that page.
"with <trait>" does an advanced search instead of going to the subset page.

Much work is done to get the highlighting right for the prefix portion of the match.